Move ShallowRenderer tests
